renaming and rebuilding extension
