--- ckanext-datagov_iauthfunctions/MANIFEST.in
@@ -1,4 +1,0 @@
-include README.rst
-include LICENSE
-include requirements.txt
-recursive-include ckanext/datagov_iauthfunctions *.html *.json *.js *.less *.css *.mo
