--- ckanext-datagov_iauthfunctions/setup.cfg
@@ -1,21 +1,0 @@
-[extract_messages]
-keywords = translate isPlural
-add_comments = TRANSLATORS:
-output_file = ckanext/datagov_iauthfunctions/i18n/ckanext-datagov_iauthfunctions.pot
-width = 80
-
-[init_catalog]
-domain = ckanext-datagov_iauthfunctions
-input_file = ckanext/datagov_iauthfunctions/i18n/ckanext-datagov_iauthfunctions.pot
-output_dir = ckanext/datagov_iauthfunctions/i18n
-
-[update_catalog]
-domain = ckanext-datagov_iauthfunctions
-input_file = ckanext/datagov_iauthfunctions/i18n/ckanext-datagov_iauthfunctions.pot
-output_dir = ckanext/datagov_iauthfunctions/i18n
-previous = true
-
-[compile_catalog]
-domain = ckanext-datagov_iauthfunctions
-directory = ckanext/datagov_iauthfunctions/i18n
-statistics = true
