+++ ckanext-datagov_inventory/.coveragerc
