+++ ckanext-datagov_inventory/.travis.yml
