+++ ckanext-datagov_inventory/MANIFEST.in
@@ -1,0 +1,4 @@
+include README.rst
+include LICENSE
+include requirements.txt
+recursive-include ckanext/datagov_inventory *.html *.json *.js *.less *.css *.mo
