+++ ckanext-datagov_inventory/ckanext/datagov_inventory/tests/test_plugin.py
@@ -1,5 +1,5 @@
 """Tests for plugin.py."""
-import ckanext.datagov_iauthfunctions.plugin as plugin
+import ckanext.datagov_inventory.plugin as plugin
 
 def test_plugin():
     pass
