+++ ckanext-datagov_inventory/setup.cfg
@@ -1,0 +1,21 @@
+[extract_messages]
+keywords = translate isPlural
+add_comments = TRANSLATORS:
+output_file = ckanext/datagov_inventory/i18n/ckanext-datagov_inventory.pot
+width = 80
+
+[init_catalog]
+domain = ckanext-datagov_inventory
+input_file = ckanext/datagov_inventory/i18n/ckanext-datagov_inventory.pot
+output_dir = ckanext/datagov_inventory/i18n
+
+[update_catalog]
+domain = ckanext-datagov_inventory
+input_file = ckanext/datagov_inventory/i18n/ckanext-datagov_inventory.pot
+output_dir = ckanext/datagov_inventory/i18n
+previous = true
+
+[compile_catalog]
+domain = ckanext-datagov_inventory
+directory = ckanext/datagov_inventory/i18n
+statistics = true
