+++ ckanext-datagov_inventory/setup.py
@@ -10,18 +10,18 @@ with open(path.join(here, 'README.rst'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
-    name='ckanext-datagov_iauthfunctions',
+    name='ckanext-datagov_inventory',
 
     # Versions should comply with PEP440.  For a discussion on single-sourcing
     # the version across setup.py and the project code, see
     # http://packaging.python.org/en/latest/tutorial.html#version
     version='0.0.1',
 
-    description='Datagov IAuthFunctions',
+    description='Datagov Inventory',
     long_description=long_description,
 
     # The project's main homepage.
-    url='https://github.com/GSA/ckanext-datagov_iauthfunctions',
+    url='https://github.com/GSA/ckanext-datagov_inventory',
 
     # Author details
     author='Data Gov',
@@ -80,7 +80,7 @@ setup(
     # pip to create the appropriate form of executable for the target platform.
     entry_points='''
         [ckan.plugins]
-        datagov_iauthfunctions=ckanext.datagov_iauthfunctions.plugin:Datagov_IauthfunctionsPlugin
+        datagov_inventory=ckanext.datagov_inventory.plugin:Datagov_IauthfunctionsPlugin
 
         [babel.extractors]
         ckan = ckan.lib.extract:extract_ckan
