+++ config/production.ini
@@ -100,7 +100,7 @@ solr_url = http://solr:8983/solr/inventory
 #       Add ``datapusher`` to enable DataPusher
 #               Add ``resource_proxy`` to enable resource proxying and get around the
 #               same origin policy
-ckan.plugins = usmetadata datajson datastore datapusher stats text_view recline_view googleanalyticsbasic dcat_usmetadata s3filestore datagov_iauthfunctions
+ckan.plugins = usmetadata datajson datastore datapusher stats text_view recline_view googleanalyticsbasic dcat_usmetadata s3filestore datagov_inventory
 
 # Define which views should be created by default
 # (plugins must be loaded in ckan.plugins)
@@ -210,7 +210,7 @@ keys = console
 keys = generic
 
 [logger_root]
-level = WARNING
+level = INFO
 handlers = console
 
 [logger_ckan]
