+++ docker-compose.yml
@@ -14,7 +14,7 @@ services:
       - ckan:/usr/lib/ckan
       - ./config:/etc/ckan
       - .:/opt/inventory-app
-      - ./ckanext-datagov_iauthfunctions:/usr/lib/ckan/src/ckanext-datagov_iauthfunctions
+      - ./ckanext-datagov_inventory:/usr/lib/ckan/src/ckanext-datagov_inventory
       # - /local/path/ckanext-some_extension/:/usr/lib/ckan/src/ckanext-some_extension
 
   datastore:
