+++ requirements.txt
@@ -6,7 +6,7 @@
 -e git+https://github.com/GSA/webob.git@ckan-patch#egg=webob
 ckanext-dcat-usmetadata~=0.2
 -e git+https://github.com/GSA/ckanext-s3filestore@datagov-v0.1.1#egg=ckanext-s3filestore
--e ./ckanext-datagov_iauthfunctions
+-e ./ckanext-datagov_inventory
 #
 # This file is autogenerated by pip-compile
 # To update, run:
